Fix kinds of build failure
Fix kinds of build failures caused by -Werror=sign-conversion,
-Werror=cast-qual, -Werror=conversion etc.
Eg:
src/offlinelogstorage/dlt_offline_logstorage.h:129:49: error: unsigned conversion from ‘int’ to ‘unsigned int’ changes value from ‘-1’ to ‘4294967295’ [-Werror=sign-conversion]
  129 | #define DLT_LOGSTORAGE_GZIP_ERROR               -1 /* error case */
      |                                                 ^
/dlt-daemon/src/offlinelogstorage/dlt_offline_logstorage.c:1303:36: note: in expansion of macro ‘DLT_LOGSTORAGE_GZIP_ERROR’
 1303 |         config->gzip_compression = DLT_LOGSTORAGE_GZIP_ERROR;

      |                                    ^~~~~~~~~~~~~~~~~~~~~~~~~

src/offlinelogstorage/dlt_offline_logstorage_behavior.c:851:16: error: conversion from ‘z_size_t’ {aka ‘long unsigned int’} to ‘int’ may change value [-Werror=conversion]
  851 |         return gzfwrite(ptr, size, nmemb, config->gzlog);
      |                ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

src/console/dlt-control-common.c:922:25: error: cast discards 'const' qualifier from pointer target type [-Werror=cast-qual]
  922 |     json_object_to_file((char*)filename, json_filter_obj);
      |

Use explicit type conversion to make the build can pass

+++ src/daemon/dlt-daemon.c
@@ -867,7 +867,7 @@ int option_file_parser(DltDaemonLocal *daemon_local)
 
                         if ((longval == MULTICAST_CONNECTION_DISABLED)
                             || (longval == MULTICAST_CONNECTION_ENABLED)) {
-                            daemon_local->UDPConnectionSetup = longval;
+                            daemon_local->UDPConnectionSetup = (int)longval;
                             printf("Option: %s=%s\n", token, value);
                         }
                         else {
@@ -884,7 +884,7 @@ int option_file_parser(DltDaemonLocal *daemon_local)
                     }
                     else if (strcmp(token, "UDPMulticastIPPort") == 0)
                     {
-                        daemon_local->UDPMulticastIPPort = strtol(value, NULL, 10);
+                        daemon_local->UDPMulticastIPPort = (int)strtol(value, NULL, 10);
                     }
 #endif
                     else if (strcmp(token, "BindAddress") == 0)
